Hydra experiment setup (#17)
Hydra experiment setup.

Added Hydra experiment setup for the Oracle and Pure Planner approaches in the ClutteredStorage2D environment.

Can run tests across multiple seeds and # of books, with outputs in a .csv file, and also in structured Hydra output /multirun folders.

run_experiments.py is the main experiment runner that follows the structyre of single_experiment.py from the python-research-starter repo
Introduced cluttered_storage.py as a benchmark wrapper for the ClutteredStorage2D environment, so that it'd be easy to instantiate a benchmark in run_experiments.py, keeping it more modular.

---------

Co-authored-by: Richard Zhou <rz3911@princeton.edu>
Co-authored-by: Tom Silver <tomssilver@gmail.com>